style(harvest): Rename HOCs directory 📝
